requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
decorator==5.2.1
Django==5.1.7
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
lxml==5.3.1
Markdown==3.8
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.2
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
zeep==4.3.1
